requirements.txt
streamlit>=1.28.1
akshare>=1.16.72
pandas>=2.1.1
plotly>=5.17.0
requests>=2.31.0
openai>=1.3.5
python-dotenv>=1.0.0
